+++ tests/test_chat_agent_deployment.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import Any
+from urllib.parse import urljoin
 
 import pytest
 from datarobot.rest import RESTClientObject
@@ -22,7 +23,9 @@ from openai import OpenAI
 def chat_client(
     pulumi_up: Any, llm_deployment_id: str, dr_client: RESTClientObject
 ) -> OpenAI:
-    deployment_chat_base_url = dr_client.endpoint + f"/deployments/{llm_deployment_id}/"
+    deployment_chat_base_url = urljoin(
+        dr_client.endpoint, f"/deployments/{llm_deployment_id}/"
+    )
 
     client = OpenAI(api_key=dr_client.token, base_url=deployment_chat_base_url)
     return client