+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improved error handling when prompting (react)
+- Use urljoin to ensure robust endpoint
 
 ## [0.1.13] - 2025-05-06